Add the ability to temporarily disable the Dota mod
This is designed to be used when there's a Dota 2 update which changes the strings files.
Users will need to update their mods for their game to play well with the Dota 2 update.
Until the mod update is ready, it can be disabled if the game appears broken.
On startup, check for a new mod version and re-enable the mod once downloaded.